refactor(daemon): split squad leader decision protocol
